Requested changes were implemented.
