+++ WordOfTheDay.Api/WordOfTheDay.Api.csproj
@@ -26,7 +26,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\WordOfTheDay.Domain\WordOfTheDay.Domain.csproj" />
-	<ProjectReference Include="..\SharedModelsLibrary\SharedModelsLibrary.csproj" />
   </ItemGroup>
 
 	<ItemGroup>
