+++ WordOfTheDay.Api/appsettings.json
@@ -8,6 +8,7 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "WordContext": "Server=(localdb)\\mssqllocaldb;Database=WordContext-03daec14-a505-48b7-8fb2-72bc85eb0fcd;Trusted_Connection=True;MultipleActiveResultSets=true"
+    "WordContext": "Server=(localdb)\\mssqllocaldb;Database=WordContext-03daec14-a505-48b7-8fb2-72bc85eb0fcd;Trusted_Connection=True;MultipleActiveResultSets=true",
+    "RabbitMQHost": "amqp://guest:guest@localhost:5672"
   }
 }
