+++ WordOfTheDay.Domain/DomainServicesConfiguration.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
+using MassTransit;
 
 namespace WordOfTheDay.Domain
 {
@@ -7,6 +8,18 @@ namespace WordOfTheDay.Domain
         public static void AddDomain(this IServiceCollection services)
         {
             services.AddScoped<IWordsServices, WordsServices>();
+        }
+        public static void AddConfiguredMassTransit(this IServiceCollection services, string host)
+        {
+            services.AddMassTransit(Configuration =>
+            {
+                Configuration.UsingRabbitMq((context, config) =>
+                {
+                    config.Host(host);
+                });
+            });
+
+            services.AddMassTransitHostedService();
         }
     }
 }
