+++ WordOfTheDay.Domain/WordOfTheDay.Domain.csproj
@@ -5,6 +5,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\SharedModelsLibrary\SharedModelsLibrary.csproj" />
     <ProjectReference Include="..\WordOfTheDay.Repository\WordOfTheDay.Repository.csproj" />
   </ItemGroup>
