+++ WordOfTheDay.Domain/WordsServices.cs
@@ -1,21 +1,24 @@
 ﻿using System;
-using System.Linq;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using WordOfTheDay.Repository.Entities;
 using WordOfTheDay.Repository;
 using WordOfTheDay.Repository.Models;
+using MassTransit;
+using SharedModelsLibrary;
 
 namespace WordOfTheDay.Domain
 {
     internal sealed class WordsServices : IWordsServices
     {
         private readonly IWordsRepository _wordsRepository;
-        public WordsServices (IWordsRepository wordsRepository)
+        private readonly IPublishEndpoint _publishEndpoint;
+        public WordsServices (IWordsRepository wordsRepository, IPublishEndpoint publishEndpoint)
         {
             _wordsRepository = wordsRepository;
+            _publishEndpoint = publishEndpoint;
         }
-        
+
         public Task<WordCount> WordOfTheDay()
         {
             var wordOfTheDay = _wordsRepository.WordOfTheDay();
@@ -29,12 +32,14 @@ namespace WordOfTheDay.Domain
 
             return closeWords;
         }
-        public Task PostWord(Word word)
+        public async Task<Task> PostWord(Word word)
         {
             word.AddTime = DateTime.Today.ToUniversalTime();
 
             word.Text = word.Text.ToLower();
             word.Email = word.Email.ToLower();
+
+            await _publishEndpoint.Publish(new WordInfo(word.Id, word.Email, word.Text, word.AddTime, word.LocationLongitude, word.LocationLatitude));
 
             return _wordsRepository.PostWord(word);
         }
